Biobakery Dockerfile for strainphlan
Although the python scripts for strainphlan are available in the default metaphlan installation, strainphlan has additional dependencies which are not installed in the yamp/metaphlan docker. Here we use the biobakery dockerfile but a slight modification was required to run in nextflow. Nexflow requires some basic tools, namely ps, that were not included on the base image.